Split testing code to more manageable pieces
